+++ src/components/layouts/DefaultLayout.tsx
@@ -1,7 +1,7 @@
 import React from "react";
-import StickyButton from "../components/Buttons/StickyButton";
-import Footer from "../components/Footer/footer";
-import Navbar from "../components/Navbar/Navbar";
+import StickyButton from "@/components/common/Buttons/StickyButton";
+import Footer from "@/components/common/Footer/footer";
+import Navbar from "@/components/common/Navbar/Navbar";
 
 type DefaultLayoutProps = {
   children: React.ReactNode;
