+++ src/components/layouts/Slider.tsx
@@ -1,8 +1,8 @@
 import { useEffect, useRef, useState } from "react";
 
 import styled from "styled-components";
-import useTouchScroll from "../Hooks/useTouchScroll";
-import { Constants } from "../constant";
+import useTouchScroll from "@/hooks/useTouchScroll";
+import { Constants } from "@/lib/constant";
 import { StyledBackArrow, StyledForwardArrow } from "./Arrow";
 
 // TODO 페이지 넘길 때 사진이 흘러가는 에니메이션 막기
