+++ src/pages/About.tsx
@@ -1,5 +1,5 @@
-import HomeAwards from "../components/Home/Info/HomeAwards";
-import HomeInfo from "../components/Home/Info/HomeInfo";
+import HomeAwards from "@/components/Home/Info/HomeAwards";
+import HomeInfo from "@/components/Home/Info/HomeInfo";
 import MyHelmet from "../helmet/MyHelmet";
 import DefaultLayout from "@/components/layouts/DefaultLayout";
 import MobileRowColLayout from "@/components/layouts/MobileRowColLayout";
