+++ src/pages/Home.tsx
@@ -1,9 +1,9 @@
 import { useEffect } from "react";
-import HomeSectionExercise from "../components/Home/Exercise/HomeSectionExercise";
-import HomeSectionInfo from "../components/Home/Info/HomeSectionInfo";
-import HomeSectionMain from "../components/Home/Main/HomeSectionMain";
-import HomeSectionMore from "../components/Home/More/HomeSectionMore";
-import HomeSectionNews from "../components/Home/News/HomeSectionNews";
+import HomeSectionExercise from "@/components/Home/Exercise/HomeSectionExercise";
+import HomeSectionInfo from "@/components/Home/Info/HomeSectionInfo";
+import HomeSectionMain from "@/components/Home/Main/HomeSectionMain";
+import HomeSectionMore from "@/components/Home/More/HomeSectionMore";
+import HomeSectionNews from "@/components/Home/News/HomeSectionNews";
 import { Constants } from "@/lib/constant";
 import MyHelmet from "../helmet/MyHelmet";
 import DefaultLayout from "@/components/layouts/DefaultLayout";
