+++ src/pages/News/NewsMobile.tsx
@@ -2,8 +2,8 @@ import { useEffect } from "react";
 import { useParams } from "react-router-dom";
 import styled from "styled-components";
 import RightArrow from "@/lib/assets/svgs/arrow_back_ios.svg";
-import MobilePhotoCard from "../../components/Photo/MobilePhotoCard";
-import Loading from "../../components/common/Skeletons/Loading";
+import MobilePhotoCard from "@/components/Photo/MobilePhotoCard";
+import Loading from "@/components/common/Skeletons/Loading";
 import { useNews } from "../../recoills/news";
 import { NewsParamsType } from "@/lib/types/NewsParamsType";
