+++ src/pages/Photo/PhotoMobile.tsx
@@ -3,8 +3,8 @@ import { useEffect } from "react";
 import { useLocation } from "react-router-dom";
 import styled from "styled-components";
 import RightArrow from "@/lib/assets/svgs/arrow_back_ios.svg";
-import MobilePhotoCard from "../../components/Photo/MobilePhotoCard";
-import Loading from "../../components/common/Skeletons/Loading";
+import MobilePhotoCard from "@/components/Photo/MobilePhotoCard";
+import Loading from "@/components/common/Skeletons/Loading";
 
 // TODO: 모바일 헤더 높이 44px 고정 상수화
