+++ src/pages/admin/AdminHomePage.tsx
@@ -1,4 +1,4 @@
-import AdminLogin from "../../components/admin/form/AdminLogin";
+import AdminLogin from "@/components/admin/form/AdminLogin";
 import MyHelmet from "../../helmet/MyHelmet";
 import DefaultLayout from "@/components/layouts/DefaultLayout";
 import SheetWrapper from "@/components/layouts/SheetWrapper";
