+++ src/pages/admin/WriteArticlePage.tsx
@@ -1,7 +1,7 @@
-import NewsForm from "../../components/admin/form/NewsForm";
-import NewsGalleryFrom from "../../components/admin/form/NewsGalleryFrom";
-import NoticeForm from "../../components/admin/form/NoticeForm";
-import TrainingLogForm from "../../components/admin/form/TrainingLogForm";
+import NewsForm from "@/components/admin/form/NewsForm";
+import NewsGalleryFrom from "@/components/admin/form/NewsGalleryFrom";
+import NoticeForm from "@/components/admin/form/NoticeForm";
+import TrainingLogForm from "@/components/admin/form/TrainingLogForm";
 import { Constants } from "@/lib/constant";
 import Title from "@/components/layouts/Title";
