+++ src/pages/admin/trainingLog/AdminTrainingLog.tsx
@@ -1,7 +1,7 @@
 import { useTrainings } from "@/recoills/tranings";
 import { Link } from "react-router-dom";
-import FormContainer from "../../../components/admin/form/FormContainer";
-import { NewArticleButton } from "../../../components/admin/form/StyledComponent/FormContainer";
+import FormContainer from "@/components/admin/form/FormContainer";
+import { NewArticleButton } from "@/components/admin/form/StyledComponent/FormContainer";
 import ListContainer from "@/components/layouts/ListContainer";
 import Row from "@/components/layouts/Row";
